@@ -1,16 +1,16 @@
 import { fromCollectionRef } from '../observable/fromRef';
-import { Observable } from 'rxjs';
-import { map, filter, scan } from 'rxjs/operators';
+import { Observable, SchedulerLike } from 'rxjs';
+import { map, scan } from 'rxjs/operators';
 
-import { Query, DocumentChangeType, DocumentChange, DocumentChangeAction, Action } from '../interfaces';
+import { Query, DocumentChangeType, DocumentChange, DocumentChangeAction } from '../interfaces';
 
 /**
  * Return a stream of document changes on a query. These results are not in sort order but in
  * order of occurence.
  * @param query
  */
-export function docChanges<T>(query: Query): Observable<DocumentChangeAction<T>[]> {
-  return fromCollectionRef(query)
+export function docChanges<T>(query: Query, scheduler?: SchedulerLike): Observable<DocumentChangeAction<T>[]> {
+  return fromCollectionRef(query, scheduler)
     .pipe(
       map(action =>
         action.payload.docChanges()
@@ -21,8 +21,8 @@ export function docChanges<T>(query: Query): Observable<DocumentChangeAction<T>[
  * Return a stream of document changes on a query. These results are in sort order.
  * @param query
  */
-export function sortedChanges<T>(query: Query, events: DocumentChangeType[]): Observable<DocumentChangeAction<T>[]> {
-  return fromCollectionRef(query)
+export function sortedChanges<T>(query: Query, events: DocumentChangeType[], scheduler?: SchedulerLike): Observable<DocumentChangeAction<T>[]> {
+  return fromCollectionRef(query, scheduler)
     .pipe(
       map(changes => changes.payload.docChanges()),
       scan((current, changes) => combineChanges(current, changes, events), []),
