@@ -1,16 +1,16 @@
 import { fromRef } from '../observable/fromRef';
-import { Observable, of, merge } from 'rxjs';
+import { Observable, of, merge, SchedulerLike } from 'rxjs';
 
-import { DatabaseQuery, ChildEvent, AngularFireAction, SnapshotAction } from '../interfaces';
+import { DatabaseQuery, ChildEvent, SnapshotAction } from '../interfaces';
 import { isNil } from '../utils';
 
 import { switchMap, distinctUntilChanged, scan } from 'rxjs/operators';
 
-export function listChanges<T=any>(ref: DatabaseQuery, events: ChildEvent[]): Observable<SnapshotAction<T>[]> {
-  return fromRef(ref, 'value', 'once').pipe(
+export function listChanges<T=any>(ref: DatabaseQuery, events: ChildEvent[], scheduler?: SchedulerLike): Observable<SnapshotAction<T>[]> {
+  return fromRef(ref, 'value', 'once', scheduler).pipe(
     switchMap(snapshotAction => {
       const childEvent$ = [of(snapshotAction)];
-      events.forEach(event => childEvent$.push(fromRef(ref, event)));
+      events.forEach(event => childEvent$.push(fromRef(ref, event, 'on', scheduler)));
       return merge(...childEvent$).pipe(scan(buildView, []))
     }),
     distinctUntilChanged()
@@ -28,8 +28,8 @@ function positionFor<T>(changes: SnapshotAction<T>[], key) {
 }
 
 function positionAfter<T>(changes: SnapshotAction<T>[], prevKey?: string) {
-  if(isNil(prevKey)) { 
-    return 0; 
+  if(isNil(prevKey)) {
+    return 0;
   } else {
     const i = positionFor(changes, prevKey);
     if( i === -1) {
@@ -41,7 +41,7 @@ function positionAfter<T>(changes: SnapshotAction<T>[], prevKey?: string) {
 }
 
 function buildView(current, action) {
-  const { payload, type, prevKey, key } = action; 
+  const { payload, type, prevKey, key } = action;
   const currentKeyPosition = positionFor(current, key);
   const afterPreviousKeyPosition = positionAfter(current, prevKey);
   switch (action.type) {
