@@ -8,6 +8,7 @@ import { AngularFireDatabase } from '../database';
 import { map } from 'rxjs/operators';
 
 export function createListReference<T=any>(query: DatabaseQuery, afDatabase: AngularFireDatabase): AngularFireList<T> {
+  const outsideAngularScheduler = afDatabase.schedulers.outsideAngular;
   return {
     query,
     update: createDataOperationMethod<Partial<T>>(query.ref, 'update'),
@@ -15,36 +16,18 @@ export function createListReference<T=any>(query: DatabaseQuery, afDatabase: Ang
     push: (data: T) => query.ref.push(data),
     remove: createRemoveMethod(query.ref),
     snapshotChanges(events?: ChildEvent[]) {
-      const snapshotChanges$ = snapshotChanges<T>(query, events);
-      return afDatabase.scheduler.keepUnstableUntilFirst(
-        afDatabase.scheduler.runOutsideAngular(
-          snapshotChanges$
-        )
-      );
+      return snapshotChanges<T>(query, events, outsideAngularScheduler).pipe(afDatabase.keepUnstableUntilFirst);
     },
     stateChanges(events?: ChildEvent[]) {
-      const stateChanges$ = stateChanges<T>(query, events);
-      return afDatabase.scheduler.keepUnstableUntilFirst(
-        afDatabase.scheduler.runOutsideAngular(
-          stateChanges$
-        )
-      );
+      return stateChanges<T>(query, events, outsideAngularScheduler).pipe(afDatabase.keepUnstableUntilFirst);
     },
     auditTrail(events?: ChildEvent[]) {
-      const auditTrail$ = auditTrail<T>(query, events)
-      return afDatabase.scheduler.keepUnstableUntilFirst(
-        afDatabase.scheduler.runOutsideAngular(
-          auditTrail$
-        )
-      );
+      return auditTrail<T>(query, events, outsideAngularScheduler).pipe(afDatabase.keepUnstableUntilFirst);
     },
     valueChanges(events?: ChildEvent[]) {
-      const snapshotChanges$ = snapshotChanges<T>(query, events);
-      return afDatabase.scheduler.keepUnstableUntilFirst(
-        afDatabase.scheduler.runOutsideAngular(
-          snapshotChanges$
-        )
-      ).pipe(
+      const snapshotChanges$ = snapshotChanges<T>(query, events, outsideAngularScheduler);
+      return snapshotChanges$.pipe(
+        afDatabase.keepUnstableUntilFirst,
         map(actions => actions.map(a => a.payload.val() as T))
       );
     }
