@@ -1,13 +1,11 @@
 import { DatabaseQuery, ChildEvent, AngularFireAction, SnapshotAction } from '../interfaces';
 import { fromRef } from '../observable/fromRef';
 import { validateEventsArray } from './utils';
-import { Observable, merge } from 'rxjs';
-
+import { SchedulerLike, Observable, merge } from 'rxjs';
 import { DatabaseSnapshot } from '../interfaces';
-import { AngularFireDatabase } from '../database';
 
-export function stateChanges<T>(query: DatabaseQuery, events?: ChildEvent[]) {
+export function stateChanges<T>(query: DatabaseQuery, events?: ChildEvent[], scheduler?: SchedulerLike) {
   events = validateEventsArray(events)!;
-  const childEvent$ = events.map(event => fromRef<T>(query, event));
+  const childEvent$ = events.map(event => fromRef<T>(query, event, 'on', scheduler));
   return merge(...childEvent$);
 }
