@@ -1,9 +1,9 @@
-import { Observable } from 'rxjs';
+import { Observable, SchedulerLike } from 'rxjs';
 import { fromRef } from '../observable/fromRef';
-import { DatabaseQuery, DatabaseSnapshot, AngularFireAction, SnapshotAction } from '../interfaces';
+import { DatabaseQuery, SnapshotAction } from '../interfaces';
 
-export function createObjectSnapshotChanges<T>(query: DatabaseQuery) {
+export function createObjectSnapshotChanges<T>(query: DatabaseQuery, scheduler?: SchedulerLike) {
   return function snapshotChanges(): Observable<SnapshotAction<T>> {
-    return fromRef(query, 'value');
+    return fromRef(query, 'value', 'on', scheduler);
   }
 }
